+++ nope/lib/views/media/list.php
@@ -1,13 +1,13 @@
 <div id="media" class="row tmpl-searchbar" ng-class="{'has-detail':selectedMedia}">
   <div class="backdrop" ng-click="closeDetail();"></div>
-  <div class="col-xs list-column" ng-class="{'has-more':metadata.next>metadata.actual}">
-    <nav class="navbar navbar-fixed-top navbar-light bg-faded offset-md-3 offset-sm-3 offset-lg-2 searchbar" nope-can="media.read">
+  <div class="col list-column" ng-class="{'has-more':metadata.next>metadata.actual}">
+    <nav class="navbar fixed-top navbar-light bg-faded offset-md-3 offset-sm-3 offset-lg-2 searchbar" nope-can="media.read">
       <form name="searchForm" ng-submit="search(q);">
         <div class="row">
-          <div class="col-xs-7">
+          <div class="col-7">
             <input type="text" class="form-control" ng-model="q.query" placeholder="Search" />
           </div>
-          <div class="col-xs" ng-hide="hideMimetypeOptions">
+          <div class="col" ng-hide="hideMimetypeOptions">
             <select class="form-control" ng-model="q.mimetype">
               <option value="">All</option>
               <option value="image/">Only images</option>
@@ -15,17 +15,17 @@
               <option value="provider">Only from providers</option>
             </select>
           </div>
-          <div class="col-xs-1">
+          <div class="col-1">
             <button type="submit" class="btn btn-secondary btn-info"><i class="fa fa-search"></i></button>
           </div>
-          <div class="col-xs" nope-can="media.create">
+          <div class="col" nope-can="media.create">
             <a href="" nope-upload-modal="onUploadDone()" accept="{{acceptedFiles}}" class="btn btn-outline-danger">Create new media <i class="fa fa-upload"></i></a>
           </div>
         </div>
       </form>
     </nav>
-    <div class="clearfix">
-      <p class="pagination-text text-xs-left text-muted pull-left" ng-show="contentsList.length"><a href="" class="text-primary" ng-click="bulkSelection=[].concat(contentsList);">Select all</a> | <a href="" class="text-primary" ng-click="bulkSelection=[];">Clear selection</a><span ng-show="bulkSelection.length">. Selected items: {{bulkSelection.length}}. With selected: <a href="" class="text-danger" nope-content-delete="deleteBulkContentOnClick(p);" ng-model="bulkSelection">delete</a>.</span></p>
+    <div class="info-text clearfix">
+      <p class="pagination-text text-xs-left text-muted pull-left" ng-show="contentsList.length"><a href="" class="text-primary" ng-click="bulkSelection=[].concat(contentsList);">Select all</a> | <a href="" class="text-primary" ng-click="bulkSelection=[];">Clear selection</a><span ng-show="bulkSelection.length">. Selected items: {{bulkSelection.length}}. With selected: <a href="" class="text-danger" nope-content-delete="deleteBulkContentOnClick(p);" ng-model="bulkSelection">delete</a> or <a href="" class="text-info" ng-click="bulkEditTags();">edit tags</a>.</span></p>
       <p ng-show="contentsList.length" class="pagination-text text-xs-right text-muted pull-right">Items {{metadata.actual===metadata.last?metadata.count:metadata.actual*metadata.rpp}} of {{metadata.count}}</p>
       <p class="pagination-text text-muted ng-cloak" ng-show="!contentsList.length">No media found<span ng-if="q.query || q.mimetype"> with filter "{{q}}"</span>.</p>
     </div>
@@ -35,7 +35,7 @@
           <div class="card-image-block">
             <nope-lazy src="p.preview.thumb"></nope-lazy>
             <div class="btn-group btn-group-sm toolbar" ng-if="!nope.isIframe">
-              <a ng-click="p.starred=!p.starred;save(p,$index);" class="btn text-white"><i class="fa" ng-class="{'fa-star-o':!p.starred,'fa-star':p.starred}"></i></a>
+              <a href="" ng-click="p.starred=!p.starred;save(p,$index);" class="btn text-white"><i class="fa" ng-class="{'fa-star-o':!p.starred,'fa-star':p.starred}"></i></a>
               <a href="" class="btn text-white" ng-click="rotate(p,90,$index);" ng-if="p.isImage"><i class="fa fa-rotate-left"></i></a>
               <a href="" class="btn text-white" ng-click="rotate(p,-90,$index);" ng-if="p.isImage"><i class="fa fa-rotate-right"></i></a>
               <a href="" nope-zoom="p" class="btn text-white" ng-if="p.isImage"><i class="fa fa-arrows-alt"></i></a>
@@ -108,7 +108,7 @@
         </div>
       </div>
     </div>
-    <nav class="navbar navbar-fixed-bottom navbar-light bg-faded offset-md-3 offset-sm-3 offset-lg-2 load-more" ng-if="metadata.next>metadata.actual">
+    <nav class="navbar fixed-bottom navbar-light bg-faded offset-md-3 offset-sm-3 offset-lg-2 load-more" ng-if="metadata.next>metadata.actual">
       <a href="" class="btn btn-block btn-outline-info" ng-click="search(q,metadata.next)">More</a>
     </nav>
   </div>