+++ nope/lib/views/modal/content-delete.php
@@ -9,7 +9,7 @@
     <p class="text-danger">This operation cannot be undone.</p>
   </nope-modal-body>
   <nope-modal-footer>
-  <a class="btn btn-default" nope-modal-close>Close</a>
-  <a class="btn btn-danger" ng-click="deleteContent();">Yes, delete</a>
+  <a href="" class="btn btn-default" nope-modal-close>Close</a>
+  <a href="" class="btn btn-danger" ng-click="deleteContent();">Yes, delete</a>
   </nope-modal-footer>
 </nope-modal>